- After finish recording show waveform instantly and process optimized waveform in background.  - Error fixes.

+++ app/src/main/java/com/dimowner/audiorecorder/audio/SoundFile.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2018 Dmitriy Ponomarenko
+ * Copyright (C) 2015 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -31,6 +31,10 @@ import java.nio.ByteOrder;
 import java.nio.ShortBuffer;
 import java.util.Arrays;
 
+/**
+ * This class taken from Ringdroid app.
+ * https://github.com/google/ringdroid
+ */
 public class SoundFile {
 
 	private File mInputFile = null;
@@ -125,7 +129,7 @@ public class SoundFile {
 
 		//SoundFile duration.
 		duration = format.getLong(MediaFormat.KEY_DURATION);
-		dpPerSec = ARApplication.getDpPerSecond(duration/1000000);
+		dpPerSec = ARApplication.getDpPerSecond((float) duration/1000000f);
 
 		MediaCodec codec = MediaCodec.createDecoderByType(format.getString(MediaFormat.KEY_MIME));
 		codec.configure(format, null, null, 0);
